requirements.txt
aiohttp
africastalking
amadeus
asgiref
certifi
channels
charset-normalizer
django
django-cors-headers
django-environ
django-extensions
djangorestframework
drf-yasg
idna
inflection
openapi-client
packaging
paystack-python
paystackapi
Pillow
python-dotenv
psycopg2-binary
pytz
PyYAML
reportlab
requests
schema
sqlparse
twilio
tzdata
uritemplate
urllib3
annotated-types
pydantic
pydantic-core
pyotp
opencv-python
pillow
python-decouple
social-auth-app-django
django-allauth
django-debug-toolbar